refactor handling of bitcoind output
